geo service controller has added

+++ src/main/java/com/khorunaliyev/kettu/dto/projection/FeatureInfo.java
@@ -2,12 +2,10 @@ package com.khorunaliyev.kettu.dto.projection;
 
 import java.time.LocalDateTime;
 
-/**
- * Projection for {@link com.khorunaliyev.kettu.entity.resources.Feature}
- */
+
 public interface FeatureInfo {
 
-    Long getId();
+    Integer getId();
 
     String getName();
 }
